refactor: mv to fake_client.go
