+++ internal/app.go
@@ -1,12 +1,9 @@
 package internal
 
 import (
-	"bytes"
 	"context"
 	"fmt"
-	"io"
 	"net/http"
-	"net/http/httputil"
 	"os"
 	"time"
 
@@ -31,27 +28,6 @@ type App struct {
 }
 
 var ErrAccessTokenNotfound = fmt.Errorf("access token is not found")
-
-type fakeTransport struct {
-	W io.Writer
-}
-
-func (t *fakeTransport) RoundTrip(req *http.Request) (*http.Response, error) {
-	b, err := httputil.DumpRequestOut(req, true)
-	if err != nil {
-		return nil, err
-	}
-
-	io.WriteString(t.W, "\nrequest:\n")
-	io.WriteString(t.W, "----------------------------------------\n")
-	t.W.Write(b)
-	io.WriteString(t.W, "\n")
-
-	return &http.Response{
-		StatusCode: 200,
-		Body:       io.NopCloser(bytes.NewBufferString(`{}`)),
-	}, nil
-}
 
 func NewApp(c *Config) (*App, error) {
 	// if c.Debug {
